+++ Core/printf/ft_putchar.c
@@ -1,6 +1,6 @@
 #include "ft_printf.h"
 
-void	ft_putchar(char c, int *cnt)
+void	ft_putchar(char c, size_t *cnt)
 {
 	write(1, &c, 1);
 	(*cnt)++;
